+++ assets/javascripts/kitten/components/molecules/cards/reward-summary-card/components/image.js
@@ -1,9 +1,10 @@
 import React from 'react'
 import classNames from 'classnames'
 
-export const Image = ({ className, ...props }) => (
+export const Image = ({ className, image, ...props }) => (
   <div
     {...props}
+    image={image}
     className={classNames('k-RewardSummaryCard__imageWrapper', className)}
   />
 )