+++ assets/javascripts/kitten/components/molecules/cards/reward-summary-card/stories.js
@@ -1,6 +1,8 @@
 import React from 'react'
 import { RewardSummaryCard } from './index'
+import classNames from 'classnames'
 import { Container, COLORS } from '../../../..'
+import { NoImageIcon } from '../../../../components/graphics/icons/no-image-icon'
 
 const argTypes = {
   href: {
@@ -76,6 +78,7 @@ export const Default = ({
   borderStyle,
   borderWidth,
   imageBorderRadius,
+  image,
 }) => {
   return (
     <RewardSummaryCard
@@ -87,7 +90,17 @@ export const Default = ({
       borderWidth={borderWidth}
     >
       <RewardSummaryCard.Image imageBorderRadius={imageBorderRadius}>
-        <img src="/kitten.jpg" alt="" />
+     
+        {!!image ? (
+          <img src="/kitten.jpg" alt="" {...image} />
+          ) : (
+          <NoImageIcon 
+            className={classNames(
+              'k-RewardSummaryCard__imageWrapper',
+              'k-u-hidden@m-down',
+            )}
+          />
+        )}
       </RewardSummaryCard.Image>
 
       <RewardSummaryCard.Title>